Extract LibDescriptors from concurrency profiler
