+++ src/SharpDetect.Profiler/CMakeLists.txt
@@ -13,4 +13,5 @@ add_subdirectory("LibIPC")
 add_subdirectory("LibProfilerCore")
 add_subdirectory("LibMetadata")
 add_subdirectory("LibIL")
+add_subdirectory("LibDescriptors")
 add_subdirectory("SharpDetect.Concurrency.Profiler")
