+++ src/SharpDetect.Profiler/LibDescriptors/ArgumentTypeDescriptor.h
