+++ src/SharpDetect.Profiler/LibDescriptors/CapturedArgumentDescriptor.h
