+++ src/SharpDetect.Profiler/LibDescriptors/CapturedValueDescriptor.h
