+++ src/SharpDetect.Profiler/LibDescriptors/CapturedValueFlags.h
