+++ src/SharpDetect.Profiler/LibDescriptors/Configuration.cpp
