+++ src/SharpDetect.Profiler/LibDescriptors/Configuration.h
