+++ src/SharpDetect.Profiler/LibDescriptors/HashingUtils.h
