+++ src/SharpDetect.Profiler/LibDescriptors/MethodDescriptor.cpp
