+++ src/SharpDetect.Profiler/LibDescriptors/MethodDescriptor.h
