+++ src/SharpDetect.Profiler/LibDescriptors/MethodInjectionDescriptor.cpp
