+++ src/SharpDetect.Profiler/LibDescriptors/MethodInjectionDescriptor.h
