+++ src/SharpDetect.Profiler/LibDescriptors/MethodRewritingDescriptor.cpp
