+++ src/SharpDetect.Profiler/LibDescriptors/MethodRewritingDescriptor.h
