+++ src/SharpDetect.Profiler/LibDescriptors/MethodSignatureDescriptor.cpp
