+++ src/SharpDetect.Profiler/LibDescriptors/MethodSignatureDescriptor.h
