+++ src/SharpDetect.Profiler/LibDescriptors/MethodVersionDescriptor.cpp
