+++ src/SharpDetect.Profiler/LibDescriptors/MethodVersionDescriptor.h
