+++ src/SharpDetect.Profiler/LibDescriptors/TypeInjectionDescriptor.cpp
