+++ src/SharpDetect.Profiler/LibDescriptors/TypeInjectionDescriptor.h
