+++ src/SharpDetect.Profiler/SharpDetect.Concurrency.Profiler/CMakeLists.txt
@@ -1,17 +1,9 @@
 set(SOURCES
-    "ArgumentTypeDescriptor.cpp"
     "CorProfiler.cpp"
-    "Configuration.cpp"
-    "dllmain.cpp"
-    "MethodDescriptor.cpp"
-    "MethodVersionDescriptor.cpp"
-    "MethodRewritingDescriptor.cpp"
-    "MethodSignatureDescriptor.cpp"
-    "MethodInjectionDescriptor.cpp"
-    "TypeInjectionDescriptor.cpp")
+    "dllmain.cpp")
 
 add_library (SharpDetect.Concurrency.Profiler SHARED ${SOURCES} "Profiler.def")
-target_link_libraries (SharpDetect.Concurrency.Profiler PRIVATE LibProfilerCore LibMetadata LibIL LibIPC)
+target_link_libraries (SharpDetect.Concurrency.Profiler PRIVATE LibProfilerCore LibMetadata LibIL LibDescriptors LibIPC)
 apply_profiler_compile_options(SharpDetect.Concurrency.Profiler)
 if(UNIX)
     target_link_options(SharpDetect.Concurrency.Profiler PRIVATE "LINKER:-z,noexecstack")
